:sparkles: feat(evolution_chain): Add evolution_chain functionality
This commit adds the functionality for the evolution_chain module in the PokeAPI project. It includes the implementation of the `evolutionChain` function, as well as related test files and documentation files. The `evolutionChain` function currently returns a hardcoded string "Hello world!" and will be further developed to retrieve the evolution chain data from the PokeAPI. This code was auto-generated by bit.dev and the technical phase question was included manually

Signed-off-by: Jonathan Stevens <jonathan@resnovas.com>