✨ feat(ci): add Metrics CI workflow
This commit introduces the `metrics.yml` workflow, which is responsible for collecting and displaying various metrics related to the project. This workflow runs on a schedule and on each push to the `master` and `main` branches. It utilizes the `lowlighter/metrics` GitHub Actions to generate and update the metrics. The metrics include information about the project's activities, community, repositories, and more. The generated metrics are displayed in a classic template format.

The addition of the `metrics.yml` workflow enhances the project's observability and provides valuable insights into its performance and engagement.

Signed-off-by: Jonathan Stevens <jonathan@resnovas.com>