:sparkles: chore(metrics.yml): add snake github contribution grid
This commit adds a step to the GitHub workflow to generate the GitHub contribution grid snake SVG. The snake SVG is generated using the `Platane/snk/svg-only@v3` action and can be found under the `dist/` directory. This addition enhances the visual representation of GitHub contributions and adds a snake-like effect to the grid.

Signed-off-by: Jonathan Stevens <jonathan@resnovas.com>