:sparkles: feat(metrics.yml): Add games directory for GitHub Snake
This commit adds a new directory called "games" to the workflow configuration file, "metrics.yml". The directory will contain the output files for the GitHub Snake game. Previously, the output files were stored in the "dist" directory.

With this change, the SVG files for the GitHub Snake game will be generated and stored in the "games" directory instead. This improves the organization of the repository and separates the game-related files from other artifacts.

Signed-off-by: Jonathan Stevens <jonathan@resnovas.com>